// packages/client/styles/prompt.ts
export enum TEMPLATE_PROMPT_COLORS {
  RED = '#E65C5C',
  ORANGE = '#E69545',
  YELLOW = '#D9D916',
  GREEN_LIGHT = '#95E545',
  GREEN_DARK = '#52CC52',
  GREEN = '#45E595',
  WATER = '#45E5E5',
  BLUE = '#5CA0E5',
  PURPLE = '#7373E5',
  LILAC = '#AC73E5',
  PINK = '#E573E5',
  DARK_PINK = '#E55CA1'
}